requirements.txt
pandas>=1.5.0
selenium>=4.0.0
tqdm>=4.64.0
openpyxl>=3.0.0
Pillow>=9.0.0
rich>=13.0.0 